requirements.txt
click==8.1.2
Flask==2.1.1
InquirerPy==0.3.3
Pillow==9.1.0
requests==2.27.1